refactor: reorganize publisher rendering
